backend/requirements.txt
twelvelabs
python-dotenv
openai
requests
fastapi
python-multipart
uvicorn